separated visualizing functions into new files and updated BUILD paths
